remove helpers.ts (#165)
* remove helpers.ts
