+++ packages/algorand-js/src/interpreter/opcode-list.ts
@@ -8,16 +8,16 @@ import { decode, encode } from "uint64be";
 
 import { TealError } from "../errors/errors";
 import { ERRORS } from "../errors/errors-list";
+import { compareArray } from "../lib/compare";
 import { MAX_CONCAT_SIZE, MAX_UINT64 } from "../lib/constants";
-import { assertLen, assertOnlyDigits, compareArray } from "../lib/helpers";
-import { convertToBuffer, convertToString, getEncoding } from "../lib/parse-data";
+import { assertLen, assertOnlyDigits, convertToBuffer, convertToString, getEncoding } from "../lib/parsing";
 import type { EncodingType, TEALStack } from "../types";
 import { Interpreter } from "./interpreter";
 import { Op } from "./opcode";
 import { txAppArg, txnSpecbyField } from "./txn";
 
 export const BIGINT0 = BigInt("0");
-const BIGINT1 = BigInt("1");
+export const BIGINT1 = BigInt("1");
 
 // Store TEAL version
 export class Pragma extends Op {
@@ -694,22 +694,14 @@ export class EqualTo extends Op {
     this.assertMinStackLen(stack, 2);
     const last = stack.pop();
     const prev = stack.pop();
-    if (typeof last === typeof prev) {
-      if (typeof last === "bigint") {
-        if (last === prev) {
-          stack.push(BIGINT1);
-        } else {
-          stack.push(BIGINT0);
-        }
-      } else {
-        if (compareArray(this.assertBytes(last), this.assertBytes(prev))) {
-          stack.push(BIGINT1);
-        } else {
-          stack.push(BIGINT0);
-        }
-      }
-    } else {
+    if (typeof last !== typeof prev) {
       throw new TealError(ERRORS.TEAL.INVALID_TYPE);
+    }
+    if (typeof last === "bigint") {
+      stack = this.pushBooleanCheck(stack, (last === prev));
+    } else {
+      stack = this.pushBooleanCheck(stack,
+        compareArray(this.assertBytes(last), this.assertBytes(prev)));
     }
   }
 }
@@ -729,22 +721,14 @@ export class NotEqualTo extends Op {
     this.assertMinStackLen(stack, 2);
     const last = stack.pop();
     const prev = stack.pop();
-    if (typeof last === typeof prev) {
-      if (typeof last === "bigint") {
-        if (last === prev) {
-          stack.push(BIGINT0);
-        } else {
-          stack.push(BIGINT1);
-        }
-      } else {
-        if (compareArray(this.assertBytes(last), this.assertBytes(prev))) {
-          stack.push(BIGINT0);
-        } else {
-          stack.push(BIGINT1);
-        }
-      }
-    } else {
+    if (typeof last !== typeof prev) {
       throw new TealError(ERRORS.TEAL.INVALID_TYPE);
+    }
+    if (typeof last === "bigint") {
+      stack = this.pushBooleanCheck(stack, last !== prev);
+    } else {
+      stack = this.pushBooleanCheck(stack,
+        !compareArray(this.assertBytes(last), this.assertBytes(prev)));
     }
   }
 }
